BackEnd/requirements.txt
# Flask y extensiones core
Flask==3.1.0
flask-cors==4.0.0
python-dotenv==1.0.0

# JWT Authentication
Flask-JWT-Extended==4.6.0

# Rate Limiting
Flask-Limiter==3.5.0

# MongoDB
pymongo==4.6.1
mongoengine==0.28.2

# Redis
redis==5.0.1

# Password hashing
bcrypt==4.1.2

# Validación de datos
marshmallow==3.20.1
email-validator==2.1.0

# Envío de emails
secure-smtplib==0.1.1

# Jobs programados
APScheduler==3.10.4

# Utilidades
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.0.0

# Desarrollo
black==23.12.1
flake8==7.0.0
autopep8==2.0.4

# WSGI server para producción
gunicorn==21.2.0

# Monitoreo y logging
python-json-logger==2.0.7

# Dependencias adicionales para CU-010 al CU-015

# Nuevas dependencias necesarias:
APScheduler>=3.10.4